lab2tema

+++ ConsoleApp3/domain/Sample.cs
@@ -1,6 +1,6 @@
 namespace ConsoleApp3.domain;
 
-public class Sample(SampleCategory sampleCategory, AgeCategory ageCategory) : Entity<long>
+public class Sample(SampleCategory sampleCategory, AgeCategory ageCategory) : Entity<int>
 {
     private SampleCategory SampleCategory { get; set; } = sampleCategory;
     private AgeCategory AgeCategory { get; set; } = ageCategory;
